Add navigator methods for movie, preview, settings fragment
